[SYCL][HIP] Marked failing tests as xfail or unsupported on hip_nvidia.
Tests that already failed for cuda/hip_amd are expected to fail on hip_nvidia.
Sometimes it appears that tests are failing on hip_nvidia for different reasons that the test is failing on hip_amd.
Error messages etc for hip_nvidia are noted in the test header in such cases.

There are cases that are failing on hip_nvidia that did not fail on hip_amd or cuda and require further investigation.
Eight tests are unexpectedly passing on hip_nvidia.

Signed-off-by: JackAKirk <jack.kirk@codeplay.com>